code organization: split fetcher from descriptor
